:truck: Extract CompartmentList component from Scenario.
