+++ frontend/src/components/Scenario/index.tsx
@@ -3,21 +3,15 @@ import {useAppDispatch, useAppSelector} from '../../store/hooks';
 import {useTheme} from '@mui/material/styles';
 import {useTranslation} from 'react-i18next';
 import {
-  selectCompartment,
   selectScenario,
   setMinMaxDates,
-  toggleCompartmentExpansion,
   toggleScenario,
 } from 'store/DataSelectionSlice';
 import {ScenarioCard} from './ScenarioCard';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
-import List from '@mui/material/List';
-import ListItemButton from '@mui/material/ListItemButton';
-import ListItemText from '@mui/material/ListItemText';
-import Typography from '@mui/material/Typography';
 import Dialog from '@mui/material/Dialog';
-import {ScrollSync, ScrollSyncPane} from 'react-scroll-sync';
+import {ScrollSync} from 'react-scroll-sync';
 import {
   useGetSimulationModelQuery,
   useGetSimulationModelsQuery,
@@ -25,11 +19,10 @@ import {
 } from '../../store/services/scenarioApi';
 import {setCompartments, setScenarios} from 'store/ScenarioSlice';
 import {dateToISOString, Dictionary} from 'util/util';
-import {useGetCaseDataSingleSimulationEntryQuery} from '../../store/services/caseDataApi';
-import {NumberFormatter} from '../../util/hooks';
 import ManageGroupDialog from '../ManageGroupDialog';
 import ConfirmDialog from '../shared/ConfirmDialog';
 import {CaseDataCard} from './CaseDataCard';
+import CompartmentList from './CompartmentList';
 
 /**
  * React Component to render the Scenario Cards Section
@@ -40,29 +33,16 @@ export default function Scenario(): JSX.Element {
   // State for the groups management dialog
   const [open, setOpen] = React.useState(false);
 
-  const {t, i18n} = useTranslation();
-  const {t: tBackend} = useTranslation('backend');
+  const {t} = useTranslation();
   const theme = useTheme();
 
   const dispatch = useAppDispatch();
   const [simulationModelKey, setSimulationModelKey] = useState<string>('unset');
-  const [compartmentValues, setCompartmentValues] = useState<Dictionary<number> | null>(null);
-
-  const {formatNumber} = NumberFormatter(i18n.language, 1, 0);
-
-  const getCompartmentValue = (compartment: string): string => {
-    if (compartmentValues && compartment in compartmentValues) {
-      return formatNumber(compartmentValues[compartment]);
-    }
-    return t('no-data');
-  };
+  const [compartmentValues] = useState<Dictionary<number> | null>(null);
 
   const scenarioList = useAppSelector((state) => state.scenarioList);
   const selectedScenario = useAppSelector((state) => state.dataSelection.scenario);
-  const selectedCompartment = useAppSelector((state) => state.dataSelection.compartment);
   const compartmentsExpanded = useAppSelector((state) => state.dataSelection.compartmentsExpanded);
-  const node = useAppSelector((state) => state.dataSelection.district.ags);
-  const startDay = useAppSelector((state) => state.dataSelection.minDate);
   const activeScenarios = useAppSelector((state) => state.dataSelection.activeScenarios);
 
   const [groupEditorUnsavedChanges, setGroupEditorUnsavedChanges] = useState(false);
@@ -73,14 +53,6 @@ export default function Scenario(): JSX.Element {
   const {data: simulationModelData} = useGetSimulationModelQuery(simulationModelKey, {
     skip: simulationModelKey === 'unset',
   });
-  const {data: caseData} = useGetCaseDataSingleSimulationEntryQuery(
-    {
-      node: node,
-      day: startDay ?? '',
-      groups: ['total'],
-    },
-    {skip: !startDay}
-  );
 
   useEffect(() => {
     if (simulationModelsData && simulationModelsData.results.length > 0) {
@@ -90,23 +62,11 @@ export default function Scenario(): JSX.Element {
   }, [simulationModelsData]);
 
   useEffect(() => {
-    if (caseData) {
-      setCompartmentValues(caseData.results[0].compartments);
-    }
-  }, [caseData]);
-
-  useEffect(() => {
     if (simulationModelData) {
       const {compartments} = simulationModelData.results;
       dispatch(setCompartments(compartments));
     }
   }, [simulationModelData, dispatch]);
-
-  useEffect(() => {
-    if (!selectedCompartment && scenarioList.compartments.length > 0) {
-      dispatch(selectCompartment(scenarioList.compartments[0]));
-    }
-  }, [dispatch, scenarioList.compartments, selectedCompartment]);
 
   useEffect(() => {
     if (scenarioListData) {
@@ -155,159 +115,7 @@ export default function Scenario(): JSX.Element {
           maxWidth: '100%',
         }}
       >
-        <Box
-          id='scenario-view-compartment-list-root'
-          sx={{
-            borderRight: `2px dashed ${theme.palette.divider}`,
-            flexGrow: 0,
-            flexShrink: 0,
-            flexBasis: '274px',
-            minHeight: '20vh',
-            maxWidth: '100%',
-            display: 'flex',
-            flexDirection: 'column',
-            marginTop: theme.spacing(3),
-            borderTop: '2px solid transparent', // invisible border for alignment with the scenario card
-            paddingBottom: 0,
-            paddingTop: theme.spacing(2),
-            paddingLeft: 0,
-            paddingRight: 0,
-          }}
-        >
-          <Box
-            id='scenario-view-compartment-list-date'
-            sx={{
-              display: 'flex',
-              justifyContent: 'space-between',
-              height: '3rem',
-              marginLeft: theme.spacing(3),
-              marginRight: 0,
-              marginBottom: theme.spacing(1),
-              paddingRight: theme.spacing(3),
-
-              // This invisible border mirrors the existing border of the scenario cards and ensures correct alignment.
-              borderTop: '2px solid transparent',
-            }}
-          >
-            <Typography
-              variant='h2'
-              sx={{
-                textAlign: 'left',
-                height: 'min-content',
-                // fontWeight: 'bold',
-                fontSize: '13pt',
-              }}
-            >
-              {t('scenario.simulation-start-day')}:
-            </Typography>
-            <Typography
-              variant='h2'
-              sx={{
-                textAlign: 'right',
-                height: 'min-content',
-                fontWeight: 'bold',
-                fontSize: '13pt',
-              }}
-            >
-              {startDay ? new Date(startDay).toLocaleDateString(i18n.language) : t('today')}
-            </Typography>
-          </Box>
-          <ScrollSyncPane group='compartments'>
-            <List
-              id='scenario-view-compartment-list'
-              dense={true}
-              disablePadding={true}
-              sx={{
-                maxHeight: compartmentsExpanded ? '248px' : 'auto',
-                overflowY: 'auto',
-              }}
-            >
-              {scenarioList.compartments.map((compartment, i) => (
-                // map all compartments to display compartment list
-                <ListItemButton
-                  key={compartment}
-                  sx={{
-                    display: compartmentsExpanded || i < 4 ? 'flex' : 'none',
-                    maxHeight: '36px', // Ensure, that emojis don't screw with the line height!
-                    padding: theme.spacing(1),
-                    paddingLeft: theme.spacing(3),
-                    paddingRight: theme.spacing(3),
-                    margin: theme.spacing(0),
-                    marginTop: theme.spacing(1),
-                    borderLeft: `2px ${
-                      selectedCompartment === compartment ? theme.palette.primary.main : 'transparent'
-                    } solid`,
-                    borderTop: `2px ${
-                      selectedCompartment === compartment ? theme.palette.background.paper : 'transparent'
-                    } solid`,
-                    borderBottom: `2px ${
-                      selectedCompartment === compartment ? theme.palette.background.paper : 'transparent'
-                    } solid`,
-                    '&.MuiListItemButton-root.Mui-selected': {
-                      backgroundColor: theme.palette.background.paper,
-                    },
-                  }}
-                  selected={selectedCompartment === compartment}
-                  onClick={() => {
-                    // dispatch new compartment name
-                    dispatch(selectCompartment(compartment));
-                  }}
-                >
-                  <ListItemText
-                    primary={tBackend(`infection-states.${compartment}`)}
-                    // disable child typography overriding this
-                    disableTypography={true}
-                    sx={{
-                      typography: 'listElement',
-                      fontWeight: selectedCompartment === compartment ? 'bold' : 'normal',
-                      flexGrow: 1,
-                      flexBasis: 100,
-                      zIndex: 20,
-                    }}
-                  />
-                  <ListItemText
-                    primary={getCompartmentValue(compartment)}
-                    // disable child typography overriding this
-                    disableTypography={true}
-                    sx={{
-                      typography: 'listElement',
-                      color:
-                        selectedCompartment === compartment ? theme.palette.text.primary : theme.palette.text.disabled,
-                      textAlign: 'right',
-                      flexGrow: 1,
-                      zIndex: 20,
-                    }}
-                  />
-                </ListItemButton>
-              ))}
-            </List>
-          </ScrollSyncPane>
-          <Button
-            variant='outlined'
-            color='primary'
-            sx={{
-              margin: theme.spacing(3),
-              marginTop: theme.spacing(4),
-              marginBottom: 0,
-              padding: theme.spacing(1),
-            }}
-            aria-label={t('scenario.more')}
-            onClick={() => {
-              dispatch(toggleCompartmentExpansion());
-              if (
-                scenarioList.compartments.findIndex((o) => {
-                  return o === selectedCompartment;
-                }) > 4
-              ) {
-                if (scenarioList.compartments.length > 0) {
-                  dispatch(selectCompartment(scenarioList.compartments[0]));
-                }
-              }
-            }}
-          >
-            {compartmentsExpanded ? t('less') : t('more')}
-          </Button>
-        </Box>
+        <CompartmentList />
         <Box
           id='scenario-view-scenario-card-list'
           sx={{
